Avoid C-style casts and some size type conversions
Aside from the syntactic cast changes, reduces some size types to 32-bits,
since they can never be so large in practice.  This eliminates some type
conversions and shaves a few bytes.

+++ include/ingen/EngineBase.hpp
@@ -46,7 +46,7 @@ public:
 	*/
 	virtual void init(double   sample_rate,
 	                  uint32_t block_length,
-	                  size_t   seq_size) = 0;
+	                  uint32_t seq_size) = 0;
 
 	/**
 	   Return true iff the engine and driver supports dynamic ports.
