+++ src/server/Driver.hpp
@@ -98,7 +98,7 @@ public:
 	virtual SampleCount block_length() const = 0;
 
 	/** Return the event buffer size in bytes */
-	virtual size_t seq_size() const = 0;
+	virtual uint32_t seq_size() const = 0;
 
 	/** Return the sample rate in Hz */
 	virtual SampleRate sample_rate() const = 0;
